split hs init tests into separate personalaccesskey and oauth2 test files
